% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioM2.R
\name{AddUnmapped}
\alias{AddUnmapped}
\title{Add unmapped probe}
\usage{
AddUnmapped(
  train = NULL,
  test = NULL,
  Unmapped_num = NULL,
  Add_FeartureSelection_Method = "wilcox.test",
  anno = NULL,
  len = NULL,
  verbose = TRUE,
  cores = 1
)
}
\arguments{
\item{train}{The input training dataset. The first column
is the label or the output. For binary classes,
0 and 1 are used to indicate the class member.}

\item{test}{The input test dataset. The first column
is the label or the output. For binary classes,
0 and 1 are used to indicate the class member.}

\item{Unmapped_num}{The number of unmapped probes.}

\item{Add_FeartureSelection_Method}{Feature selection methods. Available options are
c('cor', 'wilcox.test').}

\item{anno}{The annotation data stored in a data.frame for probe
mapping. It must have at least two columns named 'ID' and 'entrezID'.
(For details, please refer to data( data("MethylAnno") )}

\item{len}{The number of unmapped probes}

\item{verbose}{Whether to print running process information to the console}

\item{cores}{The number of cores used for computation.}
}
\value{
Matrix of unmapped probes
}
\description{
Add unmapped probe
}
