% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GO2ALLEGS_BP}
\alias{GO2ALLEGS_BP}
\title{An example about pathlistDB}
\format{
A list :
\describe{
...
}
}
\description{
An example about pathlistDB
}
\details{
A list of pathways with pathway IDs and their
corresponding genes ('entrezID' is used).
}
