% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GO_Ancestor}
\alias{GO_Ancestor}
\title{Pathways in the GO database and their Ancestor}
\format{
A data frame :
\describe{
...
}
}
\source{
From GO.db
}
\description{
Inclusion relationships between pathways
}
\details{
In the GO database, each pathway will have its own ancestor pathway.
Map pathways in GO database to about 20 common ancestor pathways.
}
