% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MethylData_Test}
\alias{MethylData_Test}
\title{An example about TrainData/TestData for methylation data}
\format{
A data frame :
\describe{
...
}
}
\description{
An example about TrainData/TestData for methylation data
MethylData_Test.
}
\details{
The first column
is the label or the output. For binary classes,
0 and 1 are used to indicate the class member.
}
