% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioM2.R
\name{PlotCorModule}
\alias{PlotCorModule}
\title{Correlalogram for Biological Differences Modules}
\usage{
PlotCorModule(
  PathwaysModule_obj = NULL,
  alpha = 0.7,
  begin = 0.2,
  end = 0.9,
  option = "C",
  family = "serif"
)
}
\arguments{
\item{PathwaysModule_obj}{Results produced by PathwaysModule()}

\item{alpha}{The alpha transparency, a number in (0,1). Detail for scale_fill_viridis()}

\item{begin}{The (corrected) hue in (0,1) at which the color map begins. Detail for scale_fill_viridis().}

\item{end}{The (corrected) hue in (0,1) at which the color map ends. Detail for scale_fill_viridis()}

\item{option}{A character string indicating the color map option to use. Detail for scale_fill_viridis()}

\item{family}{calligraphic style}
}
\value{
a ggplot object
}
\description{
Correlalogram for Biological Differences Modules
}
