% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioM2.R
\name{PlotPathInner}
\alias{PlotPathInner}
\title{Visualisation Original features that make up the pathway}
\usage{
PlotPathInner(
  data = NULL,
  pathlistDB = NULL,
  FeatureAnno = NULL,
  PathNames = NULL,
  p.adjust.method = "none",
  save_pdf = FALSE,
  alpha = 1,
  cols = NULL
)
}
\arguments{
\item{data}{The input omics data}

\item{pathlistDB}{A list of pathways with pathway IDs and their corresponding genes ('entrezID' is used).
For details, please refer to ( data("GO2ALLEGS_BP") )}

\item{FeatureAnno}{The annotation data stored in a data.frame for probe mapping.
It must have at least two columns named 'ID' and 'entrezID'.
(For details, please refer to data( data("MethylAnno") )}

\item{PathNames}{A vector.A vector containing the names of pathways}

\item{p.adjust.method}{p-value adjustment method.(holm", "hochberg", "hommel",
"bonferroni", "BH", "BY","fdr","none")}

\item{save_pdf}{Whether to save images in PDF format}

\item{alpha}{The alpha transparency, a number in (0,1).}

\item{cols}{palette (vector of colour names)}
}
\value{
a plot object
}
\description{
Visualisation Original features that make up the pathway
}
