% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioM2.R
\name{ShowModule}
\alias{ShowModule}
\title{Display biological information within each pathway module}
\usage{
ShowModule(obj = NULL, ID_Module = NULL, exact = TRUE, ancestor_anno = NULL)
}
\arguments{
\item{obj}{Results produced by PathwaysModule()}

\item{ID_Module}{ID of the diff module}

\item{exact}{Whether to divide GO pathways more accurately (work when ancestor_anno=NULL)}

\item{ancestor_anno}{Annotations for ancestral relationships (like data('GO_Ancestor') )}
}
\value{
List containing biologically specific information within the module
}
\description{
Display biological information within each pathway module
}
