% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioM2.R
\name{Stage2_FeartureSelection}
\alias{Stage2_FeartureSelection}
\title{Stage 2 Fearture Selection}
\usage{
Stage2_FeartureSelection(
  Stage2_FeartureSelection_Method = "RemoveHighcor",
  data = NULL,
  label = NULL,
  cutoff = NULL,
  preMode = NULL,
  classifier = NULL,
  verbose = TRUE,
  cores = 1
)
}
\arguments{
\item{Stage2_FeartureSelection_Method}{Feature selection methods. Available options are
c(NULL, 'cor', 'wilcox.test','cor_rank','wilcox.test_rank','RemoveHighcor', 'RemoveLinear').}

\item{data}{The input training dataset. The first column is the label.}

\item{label}{The label of dataset}

\item{cutoff}{The cutoff used for feature selection threshold. It can be any value
between 0 and 1.}

\item{preMode}{The prediction mode. "Currently only supports 'probability' for binary classification tasks."}

\item{classifier}{Learners in mlr3}

\item{verbose}{Whether to print running process information to the console}

\item{cores}{The number of cores used for computation.}
}
\value{
Column index of feature
}
\description{
Stage 2 Fearture Selection
}
\author{
Shunjie Zhang
}
