% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TransAnno}
\alias{TransAnno}
\title{An example about FeatureAnno for gene expression}
\format{
A data frame :
\describe{
...
}
}
\description{
An example about FeatureAnno for gene expression
}
\details{
The annotation data stored in a data.frame for probe
mapping. It must have at least two columns named 'ID' and 'entrezID'.
}
