% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioM2.R
\name{VisMultiModule}
\alias{VisMultiModule}
\title{Visualisation of the results of the analysis of the pathway modules}
\usage{
VisMultiModule(
  BioM2_pathways_obj = NULL,
  FindParaModule_obj = NULL,
  ShowModule_obj = NULL,
  PathwaysModule_obj = NULL,
  exact = TRUE,
  ancestor_anno = NULL,
  type_text_table = FALSE,
  text_table_theme = ttheme("mOrange"),
  volin = FALSE,
  control_label = 0,
  module = NULL,
  cols = NULL,
  n_neighbors = 8,
  spread = 1,
  min_dist = 2,
  target_weight = 0.5,
  size = 1.5,
  alpha = 1,
  ellipse = TRUE,
  ellipse.alpha = 0.2,
  theme = ggthemes::theme_base(base_family = "serif"),
  save_pdf = FALSE,
  width = 7,
  height = 7
)
}
\arguments{
\item{BioM2_pathways_obj}{Results produced by BioM2(,target='pathways')}

\item{FindParaModule_obj}{Results produced by FindParaModule()}

\item{ShowModule_obj}{Results produced by ShowModule()}

\item{PathwaysModule_obj}{Results produced by PathwaysModule()}

\item{exact}{Whether to divide GO pathways more accurately (work when ancestor_anno=NULL)}

\item{ancestor_anno}{Annotations for ancestral relationships (like data('GO_Ancestor') )}

\item{type_text_table}{Whether to display it in a table}

\item{text_table_theme}{The topic of this table.Detail for ggtexttable()}

\item{volin}{Can only be used when PathwaysModule_obj exists. ( Violin diagram )}

\item{control_label}{Can only be used when PathwaysModule_obj exists. ( Control group label )}

\item{module}{Can only be used when PathwaysModule_obj exists.( PathwaysModule ID )}

\item{cols}{palette (vector of colour names)}

\item{n_neighbors}{The size of local neighborhood (in terms of number of neighboring sample points) used for manifold approximation.
Larger values result in more global views of the manifold, while smaller values result in more local data being preserved.
In general values should be in the range 2 to 100.}

\item{spread}{The effective scale of embedded points. In combination with min_dist, this determines how clustered/clumped the embedded points are.}

\item{min_dist}{The effective minimum distance between embedded points.
Smaller values will result in a more clustered/clumped embedding where nearby points on the manifold are drawn closer together,
while larger values will result on a more even dispersal of points.
The value should be set relative to the spread value,
which determines the scale at which embedded points will be spread out.}

\item{target_weight}{Weighting factor between data topology and target topology.
A value of 0.0 weights entirely on data, a value of 1.0 weights entirely on target.
The default of 0.5 balances the weighting equally between data and target.
Only applies if y is non-NULL.}

\item{size}{Scatter plot point size}

\item{alpha}{Alpha for ellipse specifying the transparency level of fill color. Use alpha = 0 for no fill color.}

\item{ellipse}{logical value. If TRUE, draws ellipses around points.}

\item{ellipse.alpha}{Alpha for ellipse specifying the transparency level of fill color. Use alpha = 0 for no fill color.}

\item{theme}{Default:theme_base(base_family = "serif")}

\item{save_pdf}{Whether to save images in PDF format}

\item{width}{image width}

\item{height}{image height}
}
\value{
a ggplot2 object
}
\description{
Visualisation of the results of the analysis of the pathway modules
}
